Version 0.8.41
GCC, Clang and MSVC compiler use -Werror/-WX flag to treat warnings as errors.
Fix macOS deprecation on dialog file type filters.
Add cpp exceptions as a CMake option.
Add options for RTTI (broken for now), and exceptions.

+++ src/Animations/Sequence.cpp
@@ -102,15 +102,15 @@ namespace EmEn::Animations
 
 			case Variant::Type::Integer64 :
 				return Variant{static_cast< int64_t >(linearInterpolation< long double >(
-					itStart->second.value().asInteger64(),
-					itEnd->second.value().asInteger64(),
+					static_cast< long double >(itStart->second.value().asInteger64()),
+					static_cast< long double >(itEnd->second.value().asInteger64()),
 					factor
 				))};
 
 			case Variant::Type::UnsignedInteger64 :
 				return Variant{static_cast< uint64_t >(linearInterpolation< long double >(
-					itStart->second.value().asUnsignedInteger64(),
-					itEnd->second.value().asUnsignedInteger64(),
+					static_cast< long double >(itStart->second.value().asUnsignedInteger64()),
+					static_cast< long double >(itEnd->second.value().asUnsignedInteger64()),
 					factor
 				))};
 
@@ -269,15 +269,15 @@ namespace EmEn::Animations
 
 			case Variant::Type::Integer64 :
 				return Variant{static_cast< int64_t >(cosineInterpolation< long double >(
-					itStart->second.value().asInteger64(),
-					itEnd->second.value().asInteger64(),
+					static_cast< long double >(itStart->second.value().asInteger64()),
+					static_cast< long double >(itEnd->second.value().asInteger64()),
 					factor
 				))};
 
 			case Variant::Type::UnsignedInteger64 :
 				return Variant{static_cast< uint64_t >(cosineInterpolation< long double >(
-					itStart->second.value().asUnsignedInteger64(),
-					itEnd->second.value().asUnsignedInteger64(),
+					static_cast< long double >(itStart->second.value().asUnsignedInteger64()),
+					static_cast< long double >(itEnd->second.value().asUnsignedInteger64()),
 					factor
 				))};
